Create README.md
Power quality of the AC system has become a great concern due to the rapidly 
increased use of inductive loads. Electrical energy is wasted everyday due to the 
lagging power factor in the inductive loads used in industries. Hence, the urgent 
need to avoid this wastage of energy in our power system. Poor power factor 
results in poor reliability, safety problems and high energy cost. Internet of Things 
(IoT) potentially enhance the quality of life in the energy sector is used 
monitoring and controlling of electrical sites remotely. Many control methods for 
the Power Factor Correction (PFC) were proposed where different types of 
devices performing similar work were used.
This paper aims to measure and correct power factor using capacitor banks. The 
results are then displayed on LCD as well as sending the results through wi-fi 
module to web page to be viewed remotely. The profit gained after correcting the 
power factor is also calculated and sent to the web page for viewing by the user. A 
prototype of the work is built for achieving the aim of this paper.